Submission of Chapter 4, less hw4-16.

+++ lab4-6.cpp
@@ -1,5 +1,5 @@
 //Lab4-6.cpp - displays the area of a triangle
-//Created/revised by <your name> on <current date>
+//Revised by Andrew Drake on 06/28/2024
 
 #include <iostream>
 using namespace std;
@@ -10,9 +10,10 @@ int main()
 	int base = 11;
 	int height = 13;
 	double area = 0.0;
-
+	
 	//calculate and display area
-	area = base * height / 2;
+	
+	area = static_cast <double> (base) * height / 2;
 	cout << "Area: " << area << endl;
 
 	return 0;
